requirements.txt
# requirements.txt
fastapi
uvicorn
copernicusmarine
xarray
numpy
pandas
python-dotenv
